refactor: 패키지 정리
- Test 파일에 card 패키지와 player 패키지 생성
