+++ src/test/kotlin/blackjack/domain/card/CardsTest.kt
@@ -1,9 +1,5 @@
-package blackjack.domain
+package blackjack.domain.card
 
-import blackjack.domain.card.Card
-import blackjack.domain.card.CardNumber
-import blackjack.domain.card.CardShape
-import blackjack.domain.card.Cards
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
