+++ src/test/kotlin/blackjack/domain/player/ResultTest.kt
@@ -1,5 +1,6 @@
-package blackjack.domain
+package blackjack.domain.player
 
+import blackjack.domain.Result
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
